Demo implemented with CMake
In this commit we can see a full demo of the basic skeleton of the project using CMake. There are 3 threads for the server and 3 threads for the client, there is also the matches's monitor and the match's monitor as well. It is like the TPThreads but with steroids.
We can also find the CMakeLists.txt files in the root, server, client and common folder. There is a mini-guide to run them in the README.md file (TAKE A ClOSE LOOK AT IT!).