The client automatically joins the match N1 (hardcoded in creation of monitor_matches)
The client is automatically joining the match N1 for SDL2 tests purposes. What was done:
The client_main is now without a QT initial interface, it will be implemented in a new QT class that will have the client.h in it for sending and receiving commands.
In the client.cpp we are now sending an initial join commando to match N1 so the match is already started when the client starts to run. Now Facu can implement some SDL2 code for the match without the need of a QT menu.
The monitor_matches.cpp is creating a default match N1 in it's creation, because of what was said earlier (SDL2 purposes).
Lastly, the protocol tests that I was doing, are now on hold (commented) so it can be tested another important things such us SDL. I will continue with them but because it was uncompleted and we need to implement some SDL working code now, I decided to stand by, commit/push and continue with them in probably another new branch.