Tests are working with the dummy_socket!!
Finally, the dummy socket can be used. For running the tests now it must the code must be compiled by doing:
cmake .. -DTESTING=ON
make
 GTEST_COLOR=1 ctest --tests-dir build --output-on-failure -j 12
And it will run the tests properly. If the flag -DTESTING is missing, then it will run normaly, with the real socket and the tests will not be ran.
The changes that were done in this commit can be summarized in a #ifdef TESTING ..... do dummy_socket #else ... do socket