Testing is alive!!
Now we can set a macro called TESTING to true by doing cmake .. -DTESTING=ON that will use the dummy socket and enable testing. The README.md explains how to properly do it. Also the protocol.h and socket.h/.cpp where changed. The protocol because of comparisons needed for testing and the socket because of the TESTING macro.
At the moment, there are only two tests running and passing, one that checks that two protocols are different from each other and the other one that a default command's code is being sent and received correctly.