Match implemented (it is the MonitorMatch)
The Match (MonitorMatch) is now a thread and simulates a mini game that only try pushies to the other queues once in a second.
The monitor matches also changed a little bit in order to accommodate the Match's necessities.
The server also stops the Matches after joining all clients.
The server_client has now a lobby fase before starting the sender thread where it handles all the lobby logic in a function called "handle_lobby", then it starts the sender and uses the Match's queue
At the moment, the server works fine without any leaks, the client does not. now I am gonna change the server logic to be more lobby-match apart and at last, I'll fix the client (it has some issues as for example, leaks when exiting).