Protocol changed to be separated from lobby and match
Both the client and server protocol recv function have been divided in two functions, one fore the lobby and one for the server. The server uses them correctly, the client not yet (not separated in the ProtocolClient so the client is using the lobby one as the general recv from earlier).
Also, the match code was copied into a function called handle_match in the server_client.cpp