Two queues implemented in client
The client has now a queue for the lobby and another for the match.
There is a bug that if you enter a match and the next command is invalid, it gets stuck and I do not know why and where it gets stuck.
Also there are 9 frees that are not made (leaks). I do not know why I have leaks since all ADT are RAII. I should get some help from Mateo or something because I am not being able to see the leak.