Clang formatted with pre-commit (there are errors from cpplint and cppcheck). Fixed and added some tests
All project has been formatted and some errors from cpplint and cppcheck fixed (not all).
The tests added are basic ones but for now they test (almost) all the protocol I believe. The dummy_socket works fine, the error of segFault was because the function that created the protocols also created parsers but inside that function, therefore they then went out of scope and the protocol was left with garbage instead of parsers.