Merged to demo with the protocol/tests. Clang formatted but pre-commit fails (yaml errors predominate)
Everything works fine, but the client has errors that are too big to even read them all, sometimes it throws "still reachable" errors (which are not so severe) but sometimes throws real errors that I think they are probably some SDL errors. We should ask Mateo or/and Dipa about what to do with them.