GameCommand and Snapshots are now being sent, but it is not working properly yet
We now have a queue of GameCommands and Snapshots (not polymorphic yet). The server sends Snapshots and the client receives it. The client sends GameCommands and the server receives it. At the moment, the client is getting stuck after joining/creating a match. I am currently working in figuring out why.
The lobby looks that it is working fine (with Commnads) but the number_of_players is partially implemented so it does not show the actual number of players.
Also, I changed the GameCommand from the game_src folder because it is not working properly, so I created a GameCommand in common_src.
In the client, we added some public methods to push and pop from the queues (they are just wrappers). And finally, there are some comments that are there as a "guide" for the next implementations that need to be done.