Client is now single threaded when in lobby
The client is sending and receiving commands synchronously while in lobby and after joining/creating a match it throws the other two threads. The problem at the moment is that the Match in the server is not working properly (it does not start, it doesn't push to the sender's queues).