The communication with the client is working!
The Match has some bugs, it needs to be correctly stopped and the real logic is still missing, but we are on the right track!
When 'q' is pressed, the Server blocks trying to stop the match/es. Also, we are sending dummy Snapshots from the server at the moment.