Snapshots, GameCommands and Actions classes implemented and communicating
PLEASE SEE DIFFS!
The Snapshot's 'id' was changed to a char type.
The GameCommand has now more derived classes and some atts types were changed as well. It was also changed to a shared_ptr instead of unique_ptr because the try_pop(var) copies the var and a unique_ptr cannot be copied (we could change the try_pop method but I do not think that's a good idea).
The Actions derive classes aren't implemented yet, only a Base class action with a "start" send is implemented.
The protocol's base class was deleted, now each protocol is a lone class (not polymorphic). But the parser is still a polymorphic class, we should change that as well.
The dummy_socket was changed in order to use only the stack. It has not been tested yet.
Every 'id' in the game was changed to be a char type.
The monitor_match was finally removed.