Turn logic in Match added but not implemented
The turn logic is going to be handle by the game.
We should create a Turn class that has a ticks counter and a max_ticks that the game will handle. Also each GameCommand should notify the game if they are ending_turn commands (e.g. game_command->is_end_turn())