Actions classes added but not in use yet and send/recv worm implemented
The Actions are going to be polymorphic classes and need to be shared_ptr. They are not being used yet but we should start changing it right away.
The worms are now being sent and received (not tested) and the snapshots were updated as well.
Lastly, the QT lobby is now working fine, there was a case that was wrong (CASE_CREATED is the proper value). Also, there needs to be a "Start" button (like another QT box) to start the match (it should send a StartAction), because the Match needs to be started by a client from that match.