Pseudo-lobby (starting match) implemented
Now the part between the lobby (creating/joining/listing match) and the actual match is running. The creator of the match can only start the match, the other are just waiting for the map to come. The creator can send a 'Refresh' which sends the amout of players in that match and also can send a 'Start' which starts the actual match, sending to him and the other players the match (it is a synchronous part, until the map is sent, then the asynchronous part begins).
Also, the match_already_started was added (but not tested).