Maps and number of players added in join and create match
Now in the lobby, the server sends to the client the maps's names and the number of player of the joined/created match.
The Match is still receiving the map in it's constructor, which is wrong, it should receive it in the start() method.
Also, some tests where added. They test the lobby commands only (both sides). We should add some pseudo-lobby tests next