Worm's and platform's width and height are being sent in the protocol
Now the server can send the worm's and platform's width and height (a parser needs to be implemented in the protocol_client ASAP!).
Some constants where moved from gamebuilder.h to constants.h and MapNotFound was moved to custom_errors.h as well.
The yamlReader and map1 were changed because the platform's height and width do matter for the client and it is easier to add it that way instead of deducing it from the type.
The GameCommand was refactor so it has protected atts in the base class.
Finally, we are now testing almost (if not) all of what is being sent from client to server and from server to client.
Now, we must not forget to change the passing of the map to the match in the start() method instead of the constructors method.