Turns implemented
Turns in the server has been implemented. When there is only one worm, the turn is always the only worm's turn. But when there is more than one worm, the turn is change over 5 secs.
The TURN_TIME is set to be 5 * FPS * 2 (not sure why N * FPS does half the time..)
And the time passed in each step is turn_time -= it