Now the server client can have multiple worms and the client must send the worm_id in the action
The client must indicate the amount of worms that wants to have in the match (we should add a field that sets that in the Lobby before creating/joining a match because the Join will not let you add worms after the client joined that match). We also need to specify the worm_id in each action that is sent to the server.