Armies of worms implemented
We erased the worms_ids and now we have an army_id because the game is handling armies of worms. The major bug here is that the turns function can only handle even players.size() and even teams[0].size() (teams sizes). We should change the logic of the manage_turn so it can handle both even and odd values.
The client, server and protocol had been modified