Game ended and Match reaper implemented
Now the game has a way to know if the game has ended or not. It checks the amount of teams and/or that there is only one team with alive worms. If the game has ended, then the match stops it and when a new match is created, the monitor_matches kills the dead matches.
For the ending snapshot, a new attribute has been added to the snapshot that set the end_game bool to true and sets all the winning worm a state == WINNER (it only sends the winning worms, not the losers).