Installation added
Now you can run make install in the build folder which will create executables in the /bin which will let you run them from anywhere you want. Also the README.md was updated with this new commands.
Finally, the config file we have is being set as a default config, but if you want to use another one, you can create a config.yaml (with the same format as the default) wherever you want and pass the absolute path to the server so it can use that config instead of the default. E.g.
server 8080 /absolute/path/to/config.yaml